Task-2/requirements.txt
flask==3.0.0
werkzeug==3.0.1
gunicorn==21.2.0
prometheus-flask-exporter==0.23.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0